add comment documenting known issue
"known issue:

there is a max cable length the keysight will accept, depending on settings (frequency range, etc). if the user sets a length longer than that, the keysight will just measure up to max cable length. this leads to a discrepancy between what the vi and the keysight think the cable length is.

a possible solution is to read back the cable length (using visa command :CAT:DIST:STOP?) after writing it, and check if it is the expected value."